refactor: get-break-point 로직을 재사용 가능한 유틸 함수로 분리 (#208)
- 반응형 처리를 위한 break point 판단 로직을 공통 유틸로 분리하여 재사용성을 높였습니다.
- 기존에 컴포넌트 내에 중복 정의되어 있던 getBreakpoint 함수를 별도 모듈로 이동하였습니다.
